+++ src/service-override/configuration.ts
@@ -4,8 +4,7 @@ import { IEditorOverrideServices, StandaloneServices } from 'vs/editor/standalon
 import { ConfigurationService } from 'vs/platform/configuration/common/configurationService'
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration'
 import { URI } from 'vs/base/common/uri'
-import { Emitter, Event } from 'vs/base/common/event'
-import { FileChangeType, FileSystemProviderCapabilities, FileType, IFileChange, IFileSystemProviderWithFileReadWriteCapability } from 'vs/platform/files/common/files'
+import { Emitter } from 'vs/base/common/event'
 import { FileService } from 'vs/platform/files/common/fileService'
 import { ILogService } from 'vs/platform/log/common/log'
 import { Schemas } from 'vs/base/common/network'
@@ -14,54 +13,7 @@ import { TextResourceConfigurationService } from 'vs/editor/common/services/text
 import { SyncDescriptor } from 'vs/platform/instantiation/common/descriptors'
 import { IConfigurationRegistry, Extensions as ConfigurationExtensions } from 'vs/platform/configuration/common/configurationRegistry'
 import { Registry } from 'vs/platform/registry/common/platform'
-import { unsupported } from '../tools'
-
-function createConfigurationFileSystemProvider (settingsResource: URI, readConfiguration: () => string, onChange: Event<void>) {
-  const onDidFilesChange = new Emitter<IFileChange[]>()
-  onChange(() => {
-    onDidFilesChange.fire([{
-      type: FileChangeType.UPDATED,
-      resource: settingsResource
-    }])
-  })
-
-  const textEncoder = new TextEncoder()
-
-  const provider: IFileSystemProviderWithFileReadWriteCapability = {
-    capabilities: FileSystemProviderCapabilities.FileReadWrite,
-    onDidChangeCapabilities: Event.None,
-    onDidChangeFile: onDidFilesChange.event,
-    watch: function () {
-      // Ignore, the file will always be watched
-      return {
-        dispose () { }
-      }
-    },
-    stat: async (resource) => {
-      if (resource.toString() !== settingsResource.toString()) {
-        unsupported()
-      }
-      return {
-        type: FileType.File,
-        mtime: 0,
-        ctime: 0,
-        size: textEncoder.encode(readConfiguration()).length
-      }
-    },
-    mkdir: unsupported,
-    readdir: unsupported,
-    delete: unsupported,
-    rename: unsupported,
-    writeFile: unsupported,
-    readFile: async (resource) => {
-      if (resource.toString() !== settingsResource.toString()) {
-        unsupported()
-      }
-      return textEncoder.encode(readConfiguration())
-    }
-  }
-  return provider
-}
+import { createConfigurationFileSystemProvider } from './tools'
 
 let userConfigurationJson: string = '{}'
 const userConfigurationChangeEmitter = new Emitter<void>()
